Move rebuild logic to its own module
